WTA #31: Moved network data sources to separate package and added HomePageCache.
