+++ home/build.gradle.kts
@@ -6,6 +6,8 @@ dependencies {
     // Image Loading, Charts, Lottie Animations
     implementation("io.coil-kt:coil-compose:2.1.0")
     implementation("com.github.PhilJay:MPAndroidChart:v3.1.0")
+
+    implementation("androidx.datastore:datastore-preferences:1.0.0")
 }
 android {
     namespace = "com.jacob.wakatimeapp.home"
