+++ home/src/main/java/com/jacob/wakatimeapp/home/data/local/HomePageCache.kt
@@ -1,0 +1,47 @@
+package com.jacob.wakatimeapp.home.data.local
+
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.core.edit
+import androidx.datastore.preferences.core.longPreferencesKey
+import com.jacob.wakatimeapp.core.models.WeeklyStats
+import java.time.Instant
+import javax.inject.Inject
+import javax.inject.Singleton
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.catch
+import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.map
+import kotlinx.serialization.json.Json
+
+@Singleton
+class HomePageCache @Inject constructor(
+    private val dataStore: DataStore<Preferences>,
+    private val json: Json,
+) {
+
+    suspend fun getLastRequestTime(): Instant = dataStore.data.map {
+        val value = it[KEY_LAST_REQUEST_TIME]
+        value?.let(Instant::ofEpochMilli) ?: Instant.MIN
+    }
+        .catch { Instant.MIN }
+        .first()
+
+    suspend fun updateLastRequestTime(time: Instant = Instant.now()) {
+        dataStore.edit {
+            it[KEY_LAST_REQUEST_TIME] = time.toEpochMilli()
+        }
+    }
+
+    suspend fun getCachedData(): Flow<WeeklyStats> {
+        TODO()
+    }
+
+    suspend fun updateCache(weeklyStats: WeeklyStats) {
+        TODO("Not yet implemented")
+    }
+
+    companion object {
+        private val KEY_LAST_REQUEST_TIME = longPreferencesKey("KEY_LAST_REQUEST_TIME")
+    }
+}
